refactor: コードレビュー未対応の指摘事項を修正する
- テストの app.app_handle() を app.handle() に統一
- updated_ids を Vec から HashSet に変更して O(1) ルックアップに改善
- DnD ハンドルに role/aria-label/tabIndex を追加し ArrowUp/Down キーボード並べ替えを実装
- 非表示の削除ボタンに tabIndex={-1} を設定してフォーカス順から除外
- FooterButton を src/components/molecules に共通コンポーネントとして抽出し export/edit-cheatsheets 両ページで共有
