+++ src/components/molecules/index.ts
@@ -1,1 +1,2 @@
 export * from './CommandField'
+export * from './FooterButton'
